# services/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Headway API",
    "description": "Serves computed transit metrics with full lineage, the DQ resolution workflow, and the audited certification action. This API never computes a reported figure; it serves what the calculation library produced, joined to its provenance. Reported values are JSON strings (exact NUMERIC, never float). Auth: local accounts (ADR-0011); the native OIDC relying party is the next increment behind the same {sub, username, role} claim set.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/branding/logo": {
      "get": {
        "tags": [
          "branding"
        ],
        "summary": "Get Logo",
        "description": "Serve the agency logo. UNAUTHENTICATED by design \u2014 the app shell and\npublic pages show it before sign-in; the only per-caller control is the\nsame per-IP token bucket as the public open-data endpoint. The logo is\nthe least sensitive object in the system: a public image the agency\nchose to publish.",
        "operationId": "get_logo_branding_logo_get",
        "responses": {
          "200": {
            "description": "The agency logo bytes, with its stored content type.",
            "content": {
              "image/svg+xml": {},
              "image/png": {}
            }
          },
          "404": {
            "description": "No logo has been uploaded yet."
          }
        }
      },
      "post": {
        "tags": [
          "branding"
        ],
        "summary": "Upload Logo",
        "description": "Upload the agency logo (certifying official only \u2014 the same role that\nowns every other branding setting). SVG or PNG, at most 512 KiB. The\nbytes go to the object store BEFORE the settings row and audit event\ncommit together \u2014 the store-before-record ordering of ingest.",
        "operationId": "upload_logo_branding_logo_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_logo_branding_logo_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_logo_branding_logo_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_logo_branding_logo_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LogoUploadResponse": {
        "properties": {
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "bytes": {
            "type": "integer",
            "title": "Bytes"
          },
          "audit_event_id": {
            "type": "integer",
            "title": "Audit Event Id"
          }
        },
        "type": "object",
        "required": [
          "content_type",
          "bytes",
          "audit_event_id"
        ],
        "title": "LogoUploadResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}